9697: remove unused files, moved certain files from shared to either client or wep-api if not used elsewhere
